wwc2019/get_reviews_function/requirements.txt
google-cloud-firestore>=0.4.7
google-auth>=1.7.1